--- erpnext/accounts/report/straight_line/straight_line.js
@@ -1,14 +1,0 @@
-// Copyright (c) 2013, Web Notes Technologies Pvt. Ltd. and Contributors and contributors
-// For license information, please see license.txt
-
-frappe.query_reports["Straight Line"] = {
-	"filters": [
-		{
-			"fieldname":"fiscal_year",
-			"label": __("Fiscal Year"),
-			"fieldtype": "Link",
-			"options": "Fiscal Year",
-			"default": frappe.defaults.get_user_default("fiscal_year")
-		}
-	]
-}
