--- erpnext/accounts/report/straight_line/straight_line.json
@@ -1,17 +1,0 @@
-{
- "add_total_row": 0, 
- "apply_user_permissions": 1, 
- "creation": "2015-01-12 18:47:22.678276", 
- "disabled": 0, 
- "docstatus": 0, 
- "doctype": "Report", 
- "is_standard": "Yes", 
- "modified": "2015-01-12 18:47:22.678276", 
- "modified_by": "Administrator", 
- "module": "Accounts", 
- "name": "Straight Line", 
- "owner": "Administrator", 
- "ref_doctype": "Fixed Asset Account", 
- "report_name": "Straight Line", 
- "report_type": "Script Report"
-}
